requirements.txt
# requirements.txt
pandas
sqlalchemy
psycopg2-binary
supabase
pydantic-settings
fastapi
uvicorn
scikit-learn
scikit-surprise
vaderSentiment
streamlit
requests
streamlit-cookies-manager